eWallet-server/requirements.txt
asgiref==3.8.1
channels==4.2.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
PyJWT==2.10.1
python-decouple==3.8
qrcode==8.0
sqlparse==0.5.3
